+++ src/views/HomePage.vue
@@ -1,7 +1,9 @@
 <template>
   <div>
     <!-- Hero Section -->
-    <section class="bg-gradient-to-br from-primary-500 to-primary-700 text-white py-20">
+    <section
+      class="bg-gradient-to-br from-primary-500 to-primary-700 text-white py-20"
+    >
       <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 text-center">
         <h1 class="text-4xl md:text-5xl font-bold mb-6">
           실거주자 리뷰로 찾는<br />나만의 원룸
@@ -18,9 +20,7 @@
             placeholder="지역, 건물명으로 검색하세요"
             class="flex-1 px-4 py-3 text-gray-900 focus:outline-none rounded-lg"
           />
-          <button @click="handleSearch" class="btn-primary px-8">
-            검색
-          </button>
+          <button @click="handleSearch" class="btn-primary px-8">검색</button>
         </div>
       </div>
     </section>
@@ -33,31 +33,79 @@
         </h2>
         <div class="grid grid-cols-1 md:grid-cols-3 gap-8">
           <div class="text-center p-6">
-            <div class="w-16 h-16 bg-primary-100 rounded-full flex items-center justify-center mx-auto mb-4">
-              <svg class="w-8 h-8 text-primary-500" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M11.049 2.927c.3-.921 1.603-.921 1.902 0l1.519 4.674a1 1 0 00.95.69h4.915c.969 0 1.371 1.24.588 1.81l-3.976 2.888a1 1 0 00-.363 1.118l1.518 4.674c.3.922-.755 1.688-1.538 1.118l-3.976-2.888a1 1 0 00-1.176 0l-3.976 2.888c-.783.57-1.838-.197-1.538-1.118l1.518-4.674a1 1 0 00-.363-1.118l-3.976-2.888c-.784-.57-.38-1.81.588-1.81h4.914a1 1 0 00.951-.69l1.519-4.674z" />
+            <div
+              class="w-16 h-16 bg-primary-100 rounded-full flex items-center justify-center mx-auto mb-4"
+            >
+              <svg
+                class="w-8 h-8 text-primary-500"
+                fill="none"
+                stroke="currentColor"
+                viewBox="0 0 24 24"
+              >
+                <path
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  stroke-width="2"
+                  d="M11.049 2.927c.3-.921 1.603-.921 1.902 0l1.519 4.674a1 1 0 00.95.69h4.915c.969 0 1.371 1.24.588 1.81l-3.976 2.888a1 1 0 00-.363 1.118l1.518 4.674c.3.922-.755 1.688-1.538 1.118l-3.976-2.888a1 1 0 00-1.176 0l-3.976 2.888c-.783.57-1.838-.197-1.538-1.118l1.518-4.674a1 1 0 00-.363-1.118l-3.976-2.888c-.784-.57-.38-1.81.588-1.81h4.914a1 1 0 00.951-.69l1.519-4.674z"
+                />
               </svg>
             </div>
-            <h3 class="text-xl font-semibold text-gray-900 mb-2">실거주자 리뷰</h3>
-            <p class="text-gray-600">실제 거주 경험을 바탕으로 한 솔직한 리뷰를 확인하세요</p>
+            <h3 class="text-xl font-semibold text-gray-900 mb-2">
+              실거주자 리뷰
+            </h3>
+            <p class="text-gray-600">
+              실제 거주 경험을 바탕으로 한 솔직한 리뷰를 확인하세요
+            </p>
           </div>
           <div class="text-center p-6">
-            <div class="w-16 h-16 bg-primary-100 rounded-full flex items-center justify-center mx-auto mb-4">
-              <svg class="w-8 h-8 text-primary-500" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 20l-5.447-2.724A1 1 0 013 16.382V5.618a1 1 0 011.447-.894L9 7m0 13l6-3m-6 3V7m6 10l4.553 2.276A1 1 0 0021 18.382V7.618a1 1 0 00-.553-.894L15 4m0 13V4m0 0L9 7" />
+            <div
+              class="w-16 h-16 bg-primary-100 rounded-full flex items-center justify-center mx-auto mb-4"
+            >
+              <svg
+                class="w-8 h-8 text-primary-500"
+                fill="none"
+                stroke="currentColor"
+                viewBox="0 0 24 24"
+              >
+                <path
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  stroke-width="2"
+                  d="M9 20l-5.447-2.724A1 1 0 013 16.382V5.618a1 1 0 011.447-.894L9 7m0 13l6-3m-6 3V7m6 10l4.553 2.276A1 1 0 0021 18.382V7.618a1 1 0 00-.553-.894L15 4m0 13V4m0 0L9 7"
+                />
               </svg>
             </div>
-            <h3 class="text-xl font-semibold text-gray-900 mb-2">지도 기반 검색</h3>
-            <p class="text-gray-600">원하는 지역의 매물을 지도에서 한눈에 확인하세요</p>
+            <h3 class="text-xl font-semibold text-gray-900 mb-2">
+              지도 기반 검색
+            </h3>
+            <p class="text-gray-600">
+              원하는 지역의 매물을 지도에서 한눈에 확인하세요
+            </p>
           </div>
           <div class="text-center p-6">
-            <div class="w-16 h-16 bg-primary-100 rounded-full flex items-center justify-center mx-auto mb-4">
-              <svg class="w-8 h-8 text-primary-500" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z" />
+            <div
+              class="w-16 h-16 bg-primary-100 rounded-full flex items-center justify-center mx-auto mb-4"
+            >
+              <svg
+                class="w-8 h-8 text-primary-500"
+                fill="none"
+                stroke="currentColor"
+                viewBox="0 0 24 24"
+              >
+                <path
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  stroke-width="2"
+                  d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z"
+                />
               </svg>
             </div>
-            <h3 class="text-xl font-semibold text-gray-900 mb-2">실거래가 정보</h3>
-            <p class="text-gray-600">해당 건물의 과거 실거래가 정보를 확인하세요</p>
+            <h3 class="text-xl font-semibold text-gray-900 mb-2">
+              실거래가 정보
+            </h3>
+            <p class="text-gray-600">
+              해당 건물의 과거 실거래가 정보를 확인하세요
+            </p>
           </div>
         </div>
       </div>
@@ -68,7 +116,10 @@
       <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
         <div class="flex justify-between items-center mb-8">
           <h2 class="text-2xl font-bold text-gray-900">최근 등록 매물</h2>
-          <router-link to="/listings" class="text-primary-500 hover:text-primary-600 font-medium">
+          <router-link
+            to="/listings"
+            class="text-primary-500 hover:text-primary-600 font-medium"
+          >
             전체보기 &rarr;
           </router-link>
         </div>
@@ -111,37 +162,37 @@
 </template>
 
 <script>
-import { ref, computed, onMounted } from 'vue'
-import { useRouter } from 'vue-router'
-import { useListingStore } from '@/stores/listing'
-import ListingCard from '@/components/common/ListingCard.vue'
+import { ref, computed, onMounted } from "vue";
+import { useRouter } from "vue-router";
+import { useListingStore } from "@/stores/listing";
+import ListingCard from "@/components/listings/ListingCard.vue";
 
 export default {
-  name: 'HomePage',
+  name: "HomePage",
   components: {
-    ListingCard
+    ListingCard,
   },
   setup() {
-    const router = useRouter()
-    const listingStore = useListingStore()
-    const searchQuery = ref('')
+    const router = useRouter();
+    const listingStore = useListingStore();
+    const searchQuery = ref("");
 
-    const recentListings = computed(() => listingStore.listings.slice(0, 3))
+    const recentListings = computed(() => listingStore.listings.slice(0, 3));
 
     onMounted(async () => {
-      await listingStore.fetchListings()
-    })
+      await listingStore.fetchListings();
+    });
 
     function handleSearch() {
-      router.push({ path: '/listings', query: { q: searchQuery.value } })
+      router.push({ path: "/listings", query: { q: searchQuery.value } });
     }
 
     function goToListing(id) {
-      router.push(`/listings/${id}`)
+      router.push(`/listings/${id}`);
     }
 
     function toggleFavorite(id) {
-      listingStore.toggleFavorite(id)
+      listingStore.toggleFavorite(id);
     }
 
     return {
@@ -149,8 +200,8 @@ export default {
       recentListings,
       handleSearch,
       goToListing,
-      toggleFavorite
-    }
-  }
-}
+      toggleFavorite,
+    };
+  },
+};
 </script>
