+++ src/views/auth/RegisterPage.vue
